+++ cmake/qibuild/general.cmake
@@ -72,6 +72,7 @@ include("qibuild/submodule")
 include("qibuild/stage")
 include("qibuild/option")
 include("qibuild/codegen")
+include("qibuild/perf")
 
 # Find libraries from self sdk dir before everything else.
 qi_persistent_prepend_uniq(CMAKE_FIND_ROOT_PATH "${QI_SDK_DIR}")